Separate files and formatting
